Add support for Xbox One controllers with the original 2016 firmware
The original firmware for the Xbox One controllers used a slightly different report format, with most buttons bunched up together in one byte.

I've added support for this older firmware before, to SDL and the like. Adding it here makes it work great on my vita!

For reference, controllers reporting a PID of 0x02FD are the first revision with the current report format, 0x2E0 is the only PID I know that has the older one.